Order product method and other methods

+++ ConsoleApp Project/ConsoleApp Project/Models/OrderItem.cs
@@ -12,7 +12,9 @@ namespace ConsoleApp_Project.Models
         public Product Product { get; set; }
         public int Count { get; set; }
         public decimal Price { get; set; }
-        public int SubTotal { get; set; }
+        public decimal SubTotal { get; set; }
+
+
